+++ gridsome/lib/app/queue/FileProcessQueue.js
@@ -1,6 +1,5 @@
 const path = require('path')
 const fs = require('fs-extra')
-const isUrl = require('is-url')
 const { trim } = require('lodash')
 const md5File = require('md5-file/promise')
 const { forwardSlash } = require('../../utils')
@@ -45,7 +44,7 @@ class FileProcessQueue {
     }
 
     const { targetDir, pathPrefix } = this.config
-    const assetsDir = path.relative(targetDir, this.config.assetsDir)
+    const filesDir = path.relative(targetDir, this.config.filesDir)
     const relPath = path.relative(this.context, filePath)
 
     let filename = ''
@@ -55,7 +54,7 @@ class FileProcessQueue {
     } else {
       const { name, ext } = path.parse(relPath)
       let urlHash = ''
-      
+
       if (options.hash) {
         const hash = await md5File(filePath)
         urlHash = `.${!process.env.GRIDSOME_TEST ? hash : 'test'}`
@@ -65,7 +64,7 @@ class FileProcessQueue {
     }
 
     return {
-      src: forwardSlash(path.join(pathPrefix, assetsDir, 'files', filename))
+      src: forwardSlash(path.join(pathPrefix, filesDir, filename))
     }
   }
 }