+++ gridsome/lib/build.js
@@ -69,8 +69,9 @@ async function createRenderQueue ({ router, config, graphql }) {
     const { route } = router.resolve(fullPath)
     const { query } = page.pageQuery
     const routePath = trim(route.path, '/')
-    const dataPath = !routePath ? 'index.json' : `${routePath}.json`
-    const htmlOutput = path.resolve(config.targetDir, routePath, 'index.html')
+    const filePath = routePath.split('/').map(decodeURIComponent).join('/')
+    const dataPath = !routePath ? 'index.json' : `${filePath}.json`
+    const htmlOutput = path.resolve(config.targetDir, filePath, 'index.html')
     const dataOutput = path.resolve(config.cacheDir, 'data', dataPath)
 
     // TODO: remove this before v1.0
@@ -90,8 +91,9 @@ async function createRenderQueue ({ router, config, graphql }) {
     const { route } = router.resolve(node.path)
     const { query } = page.pageQuery
     const routePath = trim(route.path, '/')
-    const htmlOutput = path.resolve(config.targetDir, routePath, 'index.html')
-    const dataOutput = path.resolve(config.cacheDir, 'data', `${routePath}.json`)
+    const filePath = routePath.split('/').map(decodeURIComponent).join('/')
+    const htmlOutput = path.resolve(config.targetDir, filePath, 'index.html')
+    const dataOutput = path.resolve(config.cacheDir, 'data', `${filePath}.json`)
 
     // TODO: remove this before v1.0
     const output = path.dirname(htmlOutput)
@@ -212,7 +214,7 @@ async function processFiles (queue, { outDir }) {
   console.info(`Process files (${totalFiles} files) - ${timer(hirestime.S)}s`)
 }
 
-async function processImages (queue, { outDir, minProcessImageWidth }) {
+async function processImages (queue, { outDir, imageCacheDir, minProcessImageWidth }) {
   const timer = hirestime()
   const chunks = chunk(queue.queue, 100)
   const worker = createWorker('image-processor')
@@ -223,6 +225,7 @@ async function processImages (queue, { outDir, minProcessImageWidth }) {
       await worker.process({
         queue,
         outDir,
+        cacheDir: imageCacheDir,
         minWidth: minProcessImageWidth
       })
     } catch (err) {