+++ gridsome/lib/graphql/schema/nodes/index.js
@@ -20,6 +20,7 @@ module.exports = store => {
 
   return {
     queries,
-    connections
+    connections,
+    nodeTypes
   }
 }