+++ gridsome/lib/graphql/schema/types/file.js
@@ -1,4 +1,4 @@
-const fs = require('fs')
+const url = require('url')
 const path = require('path')
 const mime = require('mime-types')
 
@@ -11,7 +11,13 @@ exports.isFile = value => {
     const mimeType = mime.lookup(value)
     const ext = path.extname(value).toLowerCase()
 
-    if (mimeType && mimeType !== 'application/x-msdownload') {
+    if (ext.length && mimeType && mimeType !== 'application/x-msdownload') {
+      const { hostname, pathname } = url.parse(value)
+
+      if (hostname && pathname === '/') {
+        return false
+      }
+
       return !SUPPORTED_IMAGE_TYPES.includes(ext)
     }
   }
@@ -24,7 +30,7 @@ exports.fileType = {
   args: {},
   async resolve (obj, args, context, info) {
     const value = fieldResolver(obj, args, context, info)
-    
+
     if (!value) return null
 
     const result = await context.queue.add(value)