+++ gridsome/lib/graphql/schema/types/image.js
@@ -1,4 +1,3 @@
-const fs = require('fs')
 const path = require('path')
 
 const {
@@ -14,7 +13,7 @@ exports.isImage = value => {
     const ext = path.extname(value).toLowerCase()
 
     if (SUPPORTED_IMAGE_TYPES.includes(ext)) {
-      return true
+      return path.parse(value).base !== value
     }
   }