+++ gridsome/lib/server/middlewares/assets.js
@@ -2,34 +2,28 @@ const path = require('path')
 const fs = require('fs-extra')
 const { mapValues } = require('lodash')
 const { createWorker } = require('../../workers')
+const { SUPPORTED_IMAGE_TYPES } = require('../../utils/constants')
 
 module.exports = ({ context, config, queue }) => {
-  const assetsDir = path.relative(config.targetDir, config.assetsDir)
   const worker = createWorker('image-processor')
 
   return async (req, res, next) => {
-    const options = mapValues(req.query, value => {
-      return decodeURIComponent(value)
-    })
-
     const relPath = req.params[1].replace(/(%2e|\.){2}/g, '')
     const filePath = path.join(context, relPath)
+    const asset = queue.get(filePath)
 
     if (
       !filePath.startsWith(context) ||
-      !fs.existsSync(filePath)
+      !fs.existsSync(filePath) ||
+      !asset
     ) {
       return res.sendStatus(404)
     }
 
     const { ext } = path.parse(filePath)
-    let asset
-
-    try {
-      asset = await queue.add(filePath, options)
-    } catch (err) {
-      return next(err)
-    }
+    const options = mapValues(req.query, value => {
+      return decodeURIComponent(value)
+    })
 
     const serveFile = async file => {
       const buffer = await fs.readFile(file)
@@ -45,12 +39,10 @@ module.exports = ({ context, config, queue }) => {
     }
 
     try {
-      if (asset.type === 'image') {
-        const destPath = path.join(
-          config.cacheDir,
-          assetsDir,
-          asset.cacheKey + ext
-        )
+      if (SUPPORTED_IMAGE_TYPES.includes(ext)) {
+        const imageOptions = queue.images.createImageOptions(options)
+        const filename = queue.images.createFileName(filePath, imageOptions, asset.hash)
+        const destPath = path.join(config.imageCacheDir, filename)
 
         if (!fs.existsSync(destPath)) {
           await worker.processImage({