+++ gridsome/lib/utils/log.js
@@ -12,10 +12,9 @@ exports.error = function (message, tag) {
   log('error', 'red', message, tag)
 }
 
-
 function log (type, color, message, tag) {
   const formatted = color ? chalk[color](message) : message
   const fn = console[type]
-  
-  tag ? fn(tag, '>', formatted) : fn(format)
+
+  tag ? fn(tag, '>', formatted) : fn(formatted)
 }