+++ gridsome/lib/webpack/createClientConfig.js
@@ -3,7 +3,7 @@ const createBaseConfig = require('./createBaseConfig')
 
 const resolve = p => path.resolve(__dirname, p)
 
-module.exports = app => {
+module.exports = async app => {
   const isProd = process.env.NODE_ENV === 'production'
   const config = createBaseConfig(app, { isProd, isServer: false })
   const { targetDir, clientManifestPath } = app.config
@@ -39,7 +39,7 @@ module.exports = app => {
       }])
   }
 
-  app.dispatchSync('chainWebpack', null, config, {
+  await app.dispatch('chainWebpack', null, config, {
     context: app.context,
     isServer: false,
     isProd