+++ packages/source-contentful/CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.1.1"></a>
+## [0.1.1](https://github.com/gridsome/gridsome/tree/master/packages/source-contentful/compare/@gridsome/source-contentful@0.1.0...@gridsome/source-contentful@0.1.1) (2018-11-22)
+
+**Note:** Version bump only for package @gridsome/source-contentful
+
+
+
+
+
 <a name="0.1.0"></a>
 # [0.1.0](https://github.com/gridsome/gridsome/tree/master/packages/source-contentful/compare/@gridsome/source-contentful@0.0.3...@gridsome/source-contentful@0.1.0) (2018-11-11)