+++ packages/source-faker/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.2.0"></a>
+# [0.2.0](https://github.com/gridsome/gridsome/tree/master/packages/source-faker/compare/@gridsome/source-faker@0.1.0...@gridsome/source-faker@0.2.0) (2018-11-22)
+
+
+### Features
+
+* **faker:** option for custom route ([#34](https://github.com/gridsome/gridsome/tree/master/packages/source-faker/issues/34)) ([8dd69d9](https://github.com/gridsome/gridsome/tree/master/packages/source-faker/commit/8dd69d9))
+
+
+
+
+
 <a name="0.1.0"></a>
 # 0.1.0 (2018-11-11)